Simple  is Perfect
CentraBank class was rewritten
Now Rawmatrials are stored in a Dictionary that displays the reference data as a key value.

RawBankBase Class has been updated. It is no longer used for storage purposes. It is only used for data transfer from Inpector to Dictionary. The lists are cleared after the data is transferred.